Makefile: Add cache dir to the build and unit-test targets
Currently, the build and unit tests are not using a cache, thus taking
a few minutes to complete their operation.

Add a cache directory in order to speed up the build
and unit-test processes.

Add the cache directory to the `.gitignore` and `.dockerignore` files,
so it will be ignored by git and podman / docker.

Signed-off-by: Orel Misan <omisan@redhat.com>