checkup, tests: Check if VMI exists before deleting it
Currently, the existence of the VMI is not checked before
deleting the object.

Add a check if the VMI exists before deleting it, or
return an error in case it doesn't exist.

This uncovered a problem that was caused by the fact
the VMI namespace was not stored.

Store the VMI namespace upon its creation.

Based on https://github.com/kiagnose/kubevirt-dpdk-checkup/pull/30

Signed-off-by: Orel Misan <omisan@redhat.com>