Makefile: Dynamically build checkup image name
Currently, the checkup image name is hard coded.
This is problematic for developers wanting to create
this own build of the checkup and push it to their own
registry.

Use the REG and ORG variables when creating the image name.

Example usage:

```bash
ORG=myorg make build push
```

Will build and push the `quay.io/myorg/kubevirt-realtime-checkup:devel`
image.

Signed-off-by: Orel Misan <omisan@redhat.com>