oslat: New oslat client package with Run method
Introducing a package that will handle the oslat test [0], and the run
method used to run the test.
The Run command runs the oslat test inside the VMI console.
In order to make sure that the test duration does not exceed the checkup
timeout, passing the checkup ctx and wrapping the run command.
The output is checked for successful return value and then parsed for
the
maximum latency value out of the cores that are tested is returned.

the oslat command is run under the taskset command that makes sure that
it runs only on the specified cores under test.
flags used on the oslat command [0]:
- cpu-list: vCPUs under test.
- rtprio: SCHED_FIFO priority (1-99).
- duration: oslat test duration.
- workload: kind of workload tested.
- workload-mem: Size of the memory to use for the workload.

[0] https://www.mankier.com/8/oslat

Signed-off-by: Ram Lavi <ralavi@redhat.com>