tests/reporter: Introduce package and Cleanup method
The reporter package will manage the artifacts folder, cleaning it
before the test starts, and populating it with artifacts after the test
ends.
The CheckupReporter is exposing the parameters as public, in order to
introduce more parameters in later commits.
This commit introduces the package skeletin and the Cleanup method.

Signed-off-by: Ram Lavi <ralavi@redhat.com>