checkup: Change CPUs count to four
Currently the vmi-under-test is using 3 CPUs, in order to be able to run
the checkup and avoid getting the CNV-31584 [0] Jira Bug.
However, this configuration is sub-optimal, as the CPUs currently
running the Oslat are sibling to the non-isolated CPU 0.
Now that the CNV-31584 Bug is resolved, optimizing the checkup's
performance by requesting for a full core to run the Oslat test.
Setting the CPUs requested to four.

[0] https://issues.redhat.com/browse/CNV-31584

Signed-off-by: Ram Lavi <ralavi@redhat.com>