Console: Log in as root
Currently, the checkup creates a `cloud-user` user
using cloud-init, and logs in to the VMs using its
credentials.
Immediately after logging in, the checkup issues the
`sudo su` command, thus effectively running as the root user.

Logging in as root simplifies the login operation, and remove
problems that can occur when certain assumptions are made when
performing the `sudo su` step - for example:
1. Not expecting the sudo warning message
2. Not expecting to enter the user's password again.

Also remove the creation cloud-init volume, since
it is no longer needed.

Signed-off-by: Orel Misan <omisan@redhat.com>