Checkup, Setup: Create VM under test ConfigMap
In order to enable the checkup the ability to create a ConfigMap
containing a boot script to be attached to the
VM undet test, add creation of an empty ConfigMap object to the
setup flow.

The above ConfigMap is owned by the checkup, so it
will be garbage collected by K8s, in case of a failure.

Signed-off-by: Orel Misan <omisan@redhat.com>