vm-under-test/customize-vm: Manually exit script after reboot
The script currently runs a reboot, assuming the the script will stop
working immediately after that command. This assumption is flawed,
because it reality since the script runs during the boot process, the
reboot is initiated later.
Adding an "exit 0" to ensure that the script does not run after the
reboot command is set.

Signed-off-by: Ram Lavi <ralavi@redhat.com>