VM under test: Use CentOS stream 9
CentOS stream 8 had reached EOL [1] and it breaks our automation for
building the container disk image due to [2].

Our console package does not correctly handle Bracketed-paste [3],
disable it for now.

On CentOS stream 9 it is required to enable the `rt` repo
in order to install a real-time kernel.

The real-time tests package name was changed.

[1] https://blog.centos.org/2023/04/end-dates-are-coming-for-centos-stream-8-and-centos-linux-7/

[2] https://serverfault.com/questions/1161816/mirrorlist-centos-org-no-longer-resolve

[3] https://en.wikipedia.org/wiki/Bracketed-paste

Signed-off-by: Orel Misan <omisan@redhat.com>